Add a safety net for Manifold Sickness (#2446)
## About The Pull Request

Adds a safety net to Hereditary Manifold Sickness, which triggers if and
only if it progresses all the way from stage one to stage four without
_any_ visible symptoms having occurred. This safety net forces one of
the stage 4 visible symptoms to occur, which is also a damaging effect
that produces a sound effect (assuming the character has a set scream
noise).
## Why It's Good For The Game

Complete round removal should not be able to blindside you. This does
not guarantee it won't still blindside a player playing riskily with the
quirk either, getting some stage two symptoms and then the sickness
stealthing the rest of the way to stage 5 is still possible; and even
with this safety net triggering, the sickness has a very low chance of
going straight to stage 5 in the next few seconds anyways before you can
get a pill in.
It is probably not immediately obvious to the player if this safety net
occurs, as the visible effect is exactly the same as a normal symptom
(You would only be able to tell by the fact that there are no messages
in chat regarding the sickness between the last time you reset it with a
pill and the stage four symptom occurring), it simply removes the
possibility of the disease completely stealthing its way all the way to
stage 5 with no symptoms at all by forcing a symptom to occur.
## Proof Of Testing

Ran code both with stage_prob set to 100 in order to test this code for
when it should trigger, and then normally with the correct stage_prob to
make sure the safety net doesn't trigger if a symptom does occur.
Again, no visible difference with what should be normal gameplay, this
prevents an absence of something.
## Changelog
:cl: Cerami
qol: HMS / Chronic Illness no longer has the very low chance of going
all the way from stage 1 to stage 5 without any symptoms.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: Waterpig <wtryoutube@seznam.cz>