CentCom / NTC Changes by ChromeFoxxity - SPRITE CHANGES ONLY (#2662)
## About The Pull Request

To explain this slightly better than ChromeFoxxity did, this PR edits
all Centcom clothing to match the same color palette with darker greens
and brighter golds. As well, this changes the turtlenecks of Centcom to
have actual turtlenecks, rather than being simple ribbed sweaters.
**THIS DOES NOT INCLUDE THE LOADOUT CHANGES OR ATTEMPTED FORCING OF A
SWORD FOR REPRESENTITIVE.**

### It was tested by him apparently. I'm just remaking his PR
"atomized". If it doesn't work, that is on him for not checking
beforehand.

### Credit for the turtlenecks is to (https://github.com/Ebin-Halcyon).
If Halcyon has any issues with their turtlenecks being used, I will wipe
them from this whole PR and just recolor the existing turtleneck sprite.

### Other sprites that are used or edited for new clothing are from
TGstation.

## Why It's Good For The Game

To quote ChromeFoxxity:
> The uniforms however, looked a little stale, the shading was a little
off, from a dusty, green to a almost pitch black green instantly. This
gives it a more vibrant coloring, while also keeping that classic green
as the primary color. The gold has also been given more vibrancy, to
make it look more like gold instead of just yellow stripes, it may look
odd, but it fits more with the Captain's uniform, which these are based
off of, or look similar to.

So the shorthand answer is visual consistency. All of the existing
centcom uniforms use different shades of greens and golds for some
reason, rather than being consistent like a company's uniform would be.

## Proof Of Testing

Built on my end, so whatever that's good for.

![cmd_2024-12-09_15-57-51](https://github.com/user-attachments/assets/6880c9e4-28f8-493f-ba36-b7dcd7b0b869)

<details>
<summary>Sprites of what changed</summary>

_**Commander's Uniform and Skirt**_

![384160729-c9840989-6bad-4f72-bfda-52be30fce3d5](https://github.com/user-attachments/assets/aa38c01b-e51e-462d-bc99-8cf8067a57b6)

_**Official's Jumpsuit and Turtleneck**_

![384160867-5903847c-980e-403c-a471-6190d1235ca2](https://github.com/user-attachments/assets/0082a255-1046-4ad0-ae2d-142657c220cb)

_**ERT Officer's Turtleneck and Skirtleneck**_

![384160958-ab55df7f-df69-4d5f-be62-b27733252af8](https://github.com/user-attachments/assets/64aa6340-1248-4ec3-b9c2-2b0a6ec4dd80)

_**Centcom Replica Uniforms**_ (Notice the lack of belt and duller
colors on the pants/skirt)

![384161034-5997fd64-ebdc-4475-a7b1-11136faa336b](https://github.com/user-attachments/assets/acf1b109-3b3d-47a9-b3cf-8b84509afb63)

_**Additional Sprites**_

![384161137-64fdab07-32c7-474f-9f83-b66050921ec0](https://github.com/user-attachments/assets/ceb2345c-90ec-4788-ac41-7e593bab71b9)

</details>

## Changelog

:cl:
add: Adds different sprites to the CentCom uniforms, and clothing.
add: Adds a CentCom Carapace, and CentCom Gloves that aren't Naval.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>