Fixes an oversite that allows ghosts to make opfors. (#2713)
## About The Pull Request

Removes the ability for the ghost cafe to make opfors. The original
intention behind this code was to stop ghosts from making opfors, and
they all sort of shifted to going to the ghost cafe to circumvent that.
It's obviously an oversite.

## Why It's Good For The Game

The senior admin team all sort of said this was fine and it's kind of an
annoyance. The system is fine if you're in round asking to beat up the
chef, but some people do seriously go to the ghost cafe and ask for
ninja (or some other antag) every round and it's always kind of a
repeating cycle with new faces doing it.

## Proof Of Testing


![image](https://github.com/user-attachments/assets/af01bdb4-6a37-4531-9fb9-26b2c7bca41d)

## Changelog

:cl:
del: Ghost Cafe can no longer make opfors.
/:cl: