Biodome Blueshield Quarters fixes. (#2775)
## About The Pull Request

This PR fixes a few issues with the Blueshield Quarters on Biodome and
adds a few more essentials for the room.

This fixes the access for the Blueshield Quarters airlock along with
missing cable for the APC.
This also expands the room for the addition of a two computers. This
also adds: A wall camera, holopad, request console, intercom, windows
and curtains, fire door, and a newscaster.

## Why It's Good For The Game

It's usually good for an airlock for a restricted command area to not be
accessible to assistants.

## Proof Of Testing

StrongDMM:

![image](https://github.com/user-attachments/assets/b52355af-7d43-434d-929c-f0331b362601)

In a local:

![image](https://github.com/user-attachments/assets/b99bc97b-f58e-47c5-bbd8-a6991314c1f0)

## Changelog

:cl:
map: The Blueshield Quarters on Biodome is no longer free to access by
anyone. It has also been expanded.
/:cl: