Un-Skyrats Demonic Frost Miner Arena: AKA - The Mining Site / Observation Pit (#2789)
## About The Pull Request

Did you know that Skyrat made a version of the Demonic Frost Miner's
arena that was a singular z-level in 2021?
Did you know it hasn't been maintained since?
Did you know we've still been running that same edit even though Demonic
Frost Miner's arena has been updated visually by TG in 2023?

This PR fixes these Did You Knows by destroying the Skyrat variation
entirely so that TG#78826 actually loads properly. This makes the ruin
two z-levels again, a site above looking into the ice pit with entrances
all around the edge (just jump, i dare you), and makes the pit far less
of a solid flat ice plane and turns it into an actual mining site (the
ruin is literally called Mining Site, I shit you not).

_**There is a known issue from both testing and [TG
proper](https://github.com/tgstation/tgstation/issues/78870) where the
top ruin can generate without the bottom. This is due to how Multi-Z
Ruins place. The top is placed before the bottom, and if a ruin is
already placed where the bottom will go, the arena will not generate,
but the upper pit will generate. An example of this would be [if it
generates over the Hearthkin Camp](https://files.catbox.moe/skyte4.png).
I am currently looking into a fix at the moment, however it's not needed
at the time given Ice Moon progression regularly is broken with Wendigo
or the Clockwork Defender not spawning their ruins either. As well, this
does not seem to ruin the aesthetic of the top ruin itself as it still
serves as an observation point for the pit.**_
_**As such, I've decided to call the whole ruin Mining Site, but only
Observation Pit if only the top generates.**_

### _**This issue is possibly solved thanks to Waterpig. However in the
event it wasn’t and the maintainers so choose to avoid this issue, an
updated version of the Skyrat variant can be made instead matching the
bottom arena, but still having the portals to get in and out.**_

## Why It's Good For The Game

Actually keeps us on par with TG and removes an unmaintained ruin from
the code.
Changes the arena of Demonic Frost Miner so its not so stale (there's
lore and a single wall to hide behind if you're scared and dead bodies
wow)

## Proof Of Testing

<details>
<summary>Screenshots/Videos</summary>

### Successful Placement (Mining Site)

![dreamseeker_2024-12-27_05-14-38](https://github.com/user-attachments/assets/a38250f8-e905-44f5-b150-76e3aa2d19d1)

### Unsuccesful Placement (Observation Pit)

![dreamseeker_2024-12-27_05-04-31](https://github.com/user-attachments/assets/328e7a44-46f4-4679-bd6c-12857fafba38)

</details>

## Changelog

:cl:
fix: The updated arena for Demonic Frost Miner actually loads now
/:cl: