Speedmerge of Upstream CD delay on all action buttons (#2801)
## About The Pull Request

SPEEDMERGE FROM UPSTREAM TG
This PR reduces the cooldown on selecting which spell to cast using
hotkeys. This is because hotkeys have a different cooldown than using
your mouse to click the icons. I do not think there should be a reason
for using hotkeys to do the same thing as clicking to be clunky to the
point shown in videos below.

## Why It's Good For the game
Previous

https://github.com/user-attachments/assets/91814a17-c753-4c4b-b5cc-68e1be5528ac
Introduced here

https://github.com/user-attachments/assets/d9db3834-fc56-4a77-b4ef-4948c72b3334

Less clunky, easier to chain abilities. **It works on every single ui
button, not just spells btw. So modsuit modules that require targetting,
pretty much everything.**
(Huge thanks to ArturLang for the help with this)
## Changelog
:cl:
qol: Selecting which spells to cast with hotkeys, and using them in
general is faster.
/:cl: