The ACTUAL NRI PR. (#2809)
## About The Pull Request

I had some trouble with the previous PR because it turns out the files
for it had some unmerged changes from prev. bubber before the whole
TerraGov change. Anyways, remedied and all that. This renames and
changes flags of the NRI, and related items.
 Apologies for any formatting issues, this IS my first PR.
## Why It's Good For The Game

We're rid of Skyrat lore, it was well-received, and it was already
discussed between loreheads.

## Proof Of Testing

Generally had no issues, and I already loaded the sprites before on the
main server for personal use.

</details>

## Changelog
:cl: Virryi
image: Modified the NRI Flag icons.
code: Renamed a bunch of NRI-related items.
/:cl: