Adding a bunnysuit with more customization options (for trim, cuff, and breast color). This will be available from the loadout selection. (#2802)
## About The Pull Request

This adds a new bunnysuit. Currently only available via loadout
selection, the bunnysuit is patterned similarly to DimWhat's
Monkestation bunnysuits already in the code, but gives different color
customizations for its appearance (namely, for different trim/outline,
breast, and cuff colors).
## Why It's Good For The Game

Because in my opinion current bunnysuit options aren't good enough. The
main piece of the suit doesn't look great as monocolor, and the
colorization is wasted on parts of the sprite that barely reflect on the
sprite itself (the stockings are dark and transparent, while the cuff
buttons only change two pixels).
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>


![image](https://github.com/user-attachments/assets/23b6ad5e-e33d-4ae6-8bae-245c143e83ef)

![image](https://github.com/user-attachments/assets/286d8830-f236-4af9-83f6-c3a8f4d11611)

![image](https://github.com/user-attachments/assets/4e2fb109-d9c9-4241-a907-6394adefd579)

![image](https://github.com/user-attachments/assets/9720d27e-01af-4f17-8434-b05ef906b401)

</details>

## Changelog

:cl: DimWhat (original sprites), ArrisFairburne (edited sprites)
add: Added an alternate recolorable bunnysuit to the "suits" loadout
screen.
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>