Fix Biodome's SM waste and cabling so engineers stop going crazy (#2867)
## About The Pull Request

Fixes the broken waste pipe, broken emitter cabling, and broken maint
cabling around the SM/SMES in Biodome.

## Why It's Good For The Game

It was amusing watching it catch fire a few times, but it should be
fixed.

## Changelog

:cl: LT3
map: Biodome SM waste pipe is properly connected to the vent
map: Biodome SM cabling is properly connected to the emitters
/:cl: