Modifies Moon's AI core to be less cheesable. (#2878)
## About The Pull Request

Removes this turret controller that controls the core turrets.


![image](https://github.com/user-attachments/assets/5b4ad2e5-e512-4f2d-92a4-64d51d72a33c)

This also makes these walls into glass so the AI is better defended the
moment someone enters this room and so you can't cheese it as easily


![image](https://github.com/user-attachments/assets/b6d6bedd-bf5a-4bb7-8f4b-86b3e04eaceb)

## Why It's Good For The Game

The turret controller being placed only for the AI's direct access is a
deliberate mapping thing which holds true across all maps except Moon.

Also Moonstation's AI is very cheesable.

## Proof Of Testing

It works

## Changelog

:cl:
balance: Slightly adjusts Moonstation's AI core to be more defendable
for the AI.
/:cl: