Fixes the RBMK from processing when it didn't need to. (#2850)
## About The Pull Request

Fixes the RBMK creating too many radiation calls.
The strength of the radiation received is also now based on the radius
of the radiation. Higher radius radiation spikes equals stronger
radiation penetrations.
Caps the range of the RBMK's radiation from 20 to 10.

## Why It's Good For The Game

The RBMK was processing too frequently during meltdowns due to an
oversight on my part. During meltdowns, it was sending pulses every
atmos tick when this was too excessive. I think this was a copy+paste
error on my part because roughly the same code exists elsewhere. Object
processing processes every 2 seconds, so each pulse has a 2 second
cooldown per machine.

The range was also cut in half. The range was based o the define
"GAS_REACTION_MAXIMUM_RADIATION_PULSE_RANGE", which is used for tritium
gas fires. Turns out that range is 20, which is way too large. It is now
that number, but half that.


## Proof Of Testing

Untested. It's Sunday at 10:30pm. Will test tomorrow.

## Changelog

:cl: BurgerBB
fix: Fixes the RBMK creating too many radiation calls.
balance: The strength of the radiation received is also now based on the
radius of the radiation. Higher radius radiation spikes equals stronger
radiation penetrations. Caps the range of the RBMK's radiation from 20
to 10.
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>