[PORT] [TGUI] Color picker (#2670)
## About The Pull Request
Port of the color selection TGUI interface from
BeeStation/Artea-Station.
https://github.com/BeeStation/BeeStation-Hornet/pull/9417
https://github.com/Artea-Station/Artea-Station-Server/pull/525

This implementation includes a port and adaptation to React, as TG has
discontinued the use of Inferno.

If the Input: Enable TGUI parameter is disabled in the UI Tab, the old
color selection mechanism will be used instead. The new interface
supports copying the color value in the #343231 format and allows
entering a 6-character hexadecimal color code.
Additionally, a color palette for selection is available.
## Why It's Good For The Game
The TGUI color picker is more adaptable and prettier.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>


![image](https://github.com/user-attachments/assets/55f56807-7c76-41a1-8c75-7ed293d581e9)

![image](https://github.com/user-attachments/assets/7742e741-a16b-4708-8e7b-7891f426de32)

![image](https://github.com/user-attachments/assets/6e22e138-af41-447d-95e6-fb6ca7d067af)


</details>

## Changelog
:cl: itsmeow (original), RimiNosha, Phoenix404 (port)
add: Added TGUI-based color palette and replaced all usages of BYOND
color palettes with it (TGUI input preferences are still respected).
/:cl: