Adds digi sprites for loadout hos alt jumpsuit (#2915)
## About The Pull Request
Adds digi sprites for the hos alt jumpsuit "Black and Red Turtleneck"
available in the loadout menu, both regular and short sleeve version
First time doing this/making digi sprites, not sure if this is done
right, wanted to give it a go with an item I saw was missing them? Let
me know if it's not alright. I tried to mimic the spriting style of the
original clothing item as closely as possible while following the shape
of the basic digi jumpsuit (which I presume is the standard)
## Why It's Good For The Game
Missing digi sprites are not missing on purpose I assume but just
haven't been made
## Proof Of Testing

Looked fine ingame with both male and female body, was able to swap
between the casual worn variant fine as well
<details>
<summary>Screenshots/Videos</summary>


![image](https://github.com/user-attachments/assets/402d1665-172f-4b81-9a16-b38ca93a4236)

</details>