Adds the Galactic Federation Jacket (#2916)
## About The Pull Request
This adds the Galactic Federation jacket!


![image](https://github.com/user-attachments/assets/80ba0831-5db0-4d72-ac4c-79341469255a)

This is a nice little lore item that allows people to dress as their
favorite core world sailors.
## Why It's Good For The Game
This is good for the game because it adds more customization in a way
that ties in to the setting's lore. More of these items will help tie
peoples' characters together in a commonly understood storyline.

## Proof Of Testing
<details>

Proof of it working in the character creator:

![image](https://github.com/user-attachments/assets/633fe770-bd1e-45d6-9063-52c206a2a083)

Proof of it working on model:

![image](https://github.com/user-attachments/assets/f7c4fb27-7cad-4096-bb8d-19b7ac5625e6)



</details>

## Changelog
:cl: Tonadas
add: Added new Galactic Federation Jacket
add: Added Galactic Federation Jacket to character creator.
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>