Disables the Clown from being votable (#2941)
## About The Pull Request
What a wondrous idea, all events are equal weight. It would be a shame
if it was voted into rotation every. single. day. I can't play highpop
because a group of idiots decide to vote for the objectively worst
storyteller, that has zero care for anything. At least the Gamer tries
to make a coherent experience.

Admins have to deal with a ton of tickets, players have to deal with
frequent hellrounds, it's not fun for anyone.

## Why It's Good For The Game
The Clown is making my highpop players think this server is shit. It's
stressing everyone out and giving people the impression that
Bubberstation simply doesn't have good roleplay. Pick another
storyteller.

## Proof Of Testing
Compiled it

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
balance: The Clown is admin only, now.
/:cl:

---------

Co-authored-by: lessthanthree <83487515+lessthnthree@users.noreply.github.com>