[I think it works] Heretic reroll rite earlier, cheaper, Interlink and SSD protections to heretic targets! (#2758)
## About The Pull Request
Hello Chat. I have no idea how to test this on a private server, have no
clue if we already had these protections and don't know if the freeze
will even end in a month. What I DO know is that heretic has a lot of
frustration with it and that I want to do my part! So let me first say,
if the dorms and ssd part is unneeded/broken and a better coder says
it's easier to keep the first part than fix my spaghetti, sure, I can
just commit that out, we can go with just the easier rerolls.

Now let me argue my case:
## Why It's Good For The Game
Be a heretic. Roll 5 targets. Doesn't matter when. Someone's at the bar?
Oh welp it's either convince them away, success, or fail at that and now
you gotta stalk them out or just go guns blazing into 5 people. None of
the 6 people over there want to go through all that trouble to do the
thing they want to do. That includes YOU. And depending on the type of
RP there's protections.

Okay, so this is why heretics have a REROLL. Awesome, problem solved?
Well... Its far down the skill tree... Now you gotta wait/gather rifts
or try to get sacrifices from other people, which are guaranteed to be
at least 1 secoff at least 1 head (earlier on? No way you do this!), 1
coworker, 1 random person (Plus there's Mr. Busy). That seems like 4
more chances, but really it's 2 more.

And well, if that person's in a very open job like botany, chef, bar,
then you are going to cause a slaughter trying to get this one person.
(You cannot convince everyone, it's normal).

So when do we get the reroll?...

3 steps away from ascencion...

So what do we do? Well if you could consult the graph:

![image](https://github.com/user-attachments/assets/312392b5-dcd3-4587-84ea-e914ac5a6985)


**We let people reroll earlier on, because neither the sacrificer nor
the sacrificee wants to be involved. And there are people who WANT a
fight, Roleplay, or are idly wandering halls.**

Finally, we change the price. Whoah. Whoah whoah, ssalty, why the price
change? Well, very simply:
Do you know how much of an effort it is to start a garden while on the
run for PACTS WITH DARK GODS?

Very. But it's not as suspicious as carrying a flower that nobody but
heretics touch. Nobody. Harebells, unlike poppies, have zero use. I
don't _Want_ to argue about Back In The Day, but back in the day when
this got first introduced (we are talking 3 heretic paths, gib on
sacrifice) there'd be arrests and suspicions over someone just taking
the damn flowers!! And since then? Nothing has changed about this
mechanically. To me? This looks like:

"The antag HAS to steal an item that only they ever would touch EN
MASSE, or grow them - contradicting their playstyle to sit down and
watch paint dry/in case they like stealth, then having to do a
meaningless chore of getting the funny flower, Or have to put in energy
to get someone who doesn't want to engage _at all_ into antag stuff...
When there are better options availibl! Or maybe they just straight up
are stuck with ONE less sacrifice option because well unlucky, your
targets are ssd or ERP'ing."


If you would consult the chart
![consult the
chart](https://github.com/user-attachments/assets/418d1a28-c624-4b33-b5da-afda17006e4d)


## Proof Of Testing


![hereticced](https://github.com/user-attachments/assets/6c25ba36-5aaa-4b5f-89f0-93df9ef89aac)

</details>

## Changelog

:cl:
qol: All Mansus Worshippers can use the Relentless Heartbeat (Reroll
targets) ritual earlier by two steps, so that they gain access alongside
the ritual of knowledge.
qol: The mansus no longer demands harebells for the Relentless
Heartbeat, however a heart is used in it's place.
fix: The Grail was very adamant on keeping people making love in their
chambers out of the heretic's target list when it is made or rerolled.
/:cl: