Fixes pirates round event so they spawn in story teller (#2982)
## About The Pull Request
So it turns out that the list of possible pirates was only ever
populated during the dynamic game mode procs to run the event, since we
use story teller and not dynamic, that means the list was empty. Empty
list means no pirates. This PR makes it so that if the list is empty (it
is only ever populated when an admin uses the trigger-event verb to
manually spawn pirates) it will use the combined global lists of heavy
and light pirates, pick one, and fire it.
## Why It's Good For The Game
this antag been broke for like a year and 3 months
## Proof Of Testing

![image](https://github.com/user-attachments/assets/a249ee38-fb31-4527-88a6-a4c74c2bc931)
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
fix: pirates spawn again
/:cl: