Adds CE Screws Plushie to LustWish vendor (#2979)
Adds the Chief Screws Plush to the LustWish vendor
## About The Pull Request
Makes it so the Chief Engineering Screws Plushie can be purchasable at
the LustWish Vendor as a Toy, as it has inherently NSFW qualities.
## Why It's Good For The Game
It makes a well-sprited item actually accessible by in-game means rather
than being solely admin-bus'd in.
## Proof Of Testing
It's not tested. It's a singular line of coded added to the LustWish
vendor's inventory.
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
-Chief Engineer Screws Plush can now be purchased from the 'Premium'
section of the LustWish vendor.
:cl: Rathen Ivanov
add: You can now purchase your very own CE Screws Plushie from the
LustWish Vendor.
/:cl: