Reduces bloodsucker weight by 3 (#2961)
## About The Pull Request
Bloodsucker rolls pretty frequently, and has the same weight as traitor,
despite being a significantly more impactful antagonist. This reduces
its weight by 3, and changes the minimum number of players before it can
be spawned to 30.

## Why It's Good For The Game
Bloodsuckers roll the most out of any antagonist, and they're still a
pretty significant threat. This should help even it out.

## Proof Of Testing
It compiles

<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog

:cl: ReturnToZender
balance: Reduces the weight of Bloodsuckers by 3, to make it notably
different from traitors
/:cl: