Security BattleDress Fix + Two new VARIANTS! (#2993)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fixes the blue security battledress which previously had a bugged sprite
by adding the tag female_sprite_flag to remove the akward invisible
pixel it had (thanks waterpig for finding this)

Additionally this PR adds two new variants of the Battledress!

+ RedSec Battledress, available in the multisec vendor!
+ HoS Battledress, available in the command vendor and the HoS Garment
bag!

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Fixes a silly bug with the battledress and adds more variety!

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
<details>
<summary>Screenshots/Videos</summary>


![image](https://github.com/user-attachments/assets/4d3b362b-20a0-4769-a869-29747c8b9d46)

![image](https://github.com/user-attachments/assets/9c43d903-d535-4c3b-9037-a8a4abb33b6a)

![image](https://github.com/user-attachments/assets/36557542-b2db-4ca6-821f-664bd641aabc)

![image](https://github.com/user-attachments/assets/b1c09ee1-6784-4193-97b3-42f79cae23cc)

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Fixed Security Battledress sprite
add: Added Red Security Battledress and Head of Security Battledress
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>