Nerfs Entombed (#3026)
## About The Pull Request

Sets the fused MODsuit's defence values to be the same as the civilian
suit, removes its free insulated gloves and makes it unable to be taken
with Bad Back for the free +8 points. Additionally bumps the Mod
Complexity of the Entombed suit to 13, it will functionally be 10 since
the default storage module the entombed suits has consumes 3, and 10
complexity sets it in line with other suits.
## Why It's Good For The Game

The entombed quirk provides a comical amount of benefits for a so called
neutral quirk, including an abusable interaction with the Bad Back Quirk
that allows the player to ignore all the downsides of Bad Back and get a
free +8 points as a result.
## Proof Of Testing

Works on literally everyone else's machine. Glory to Dream-Maker. Tested
and maintainer approved.
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
balance: Made entombed not insulated, lowered its defences and is unable
to be taken with Bad Back.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>