Mini fixes to Box Station (#3029)
## About The Pull Request

This update was held back in preparation for the addition of the Torment
Nexus, but with that project on hold, this update will still be
requested as it includes essential improvements to Box. Following
feedback for Box Station, these changes include multiple fixes to
multiple hallways and its light fixtures along with the extension of
security's fore maintenance.

See the changelog for the full changes.

## Why It's Good For The Game

In addition to certain spots having their light fixtures modified/added
to improve visibility in certain areas; The largest change this request
brings is an APC is the primary hallway leading into the brig. Therefore
the area no longer infinitely powered somehow.

## Proof Of Testing

tested on my local. it works. if mapping check issues arrive, I will fix
it asap.

</details>

## Changelog

:cl:
map: In Box Station, a couple more light fixtures in the main halls were
added.
map: In Box Station, an APC in the hallway before the brig was added. It
no longer has infinite power.
map: In Box Station, a maintenance entrance in the hallway before the
brig was added. This slightly squishes the detective's office.
/:cl: