Orefruit harvesting producing sheets (#2800)
## About The Pull Request
This PR makes it so that orefruit when harvested produces cores as
sheets, instead of a grown item. I've added stack sprites too, so you
can get more of a visual update on stacks. This doesn't affect the
tedium of peeling them, you still have to do that by hand, this just
makes it so that the resulting cores can be loaded into protolathes,
machines and silos.
## Why It's Good For The Game

It was tedious for many to produce a recycler layout or autolathe to
actually make use of the ores.

## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>


![image](https://github.com/user-attachments/assets/81eb40fe-1ad5-43c3-87dc-e94ab5ce042a)


![image](https://github.com/user-attachments/assets/bc88f7bd-5481-4595-9226-f3ca33162021)


</details>

## Changelog
:cl:
qol: made orefruits produce sheets as cores, instead of lame objects,
letting them be inserted into a silo.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>