Lizard Gas Atmos Fix/I.D transferred to Modular File (#3049)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
The Lizard Gas Ghost roles use Omni pressure tanks, these are ass in
game and don't have layers properly coded in. Also the loop is currently
not hooked up. Also the Liquid plasma tank network has been changed to
have an exterior store-house, because immediately upon people trying to
fix atmos- sparks from the RPD caught the tank on fire and torched the
entire backroom, maintenance room, and manager's office in a roaring
plasma/trit fire 😔

p.s the Lizard Gas I.D was added to a non-modular file on accident, this
PR switches it to modular.

The mapsize was not altered to accommodate these changes overall and it
remains a 25x25 Ghost Role ruin (as it's always been.)

TL;DR:

-Atmos Pipes are colored Red and Blue for Scrubbers and Air Supply, as
God intended
-Exterior Scrubber added to suck Oxygen out of the air and store it in
the network/a connected Atmos Oxygen Cannister
-Mapping Error corrected, interior vent in the central room was set to
layer 4, and not hooked up.
-Exterior storage closet added for liquid plasma supply following
ridiculous flammability INSIDE of the structure (risks torching the
ENTIRE gas station because of a single spark.)
-Moves Lizard Gas I.D `item` to Modular folder, is currently (and
unnecessarily) in Non-Modular folder without being noted. No bueno.
-Grants access to alter their own air alarm
-Additional Air Scrubber and Vent added to Maintenance backroom
-Additional Air Scrubber and Vent added to Bathroom
-Moves Oxygen Atmos can inside instead of outback, pretty sure the heat
of Lavaland damages it/causes it to leak.

Edit: 
-Edited the Note that formerly mentioned the omni pressure tanks, as
manually fixing the atmos is no longer required as they have been
removed!
-Added Connector port for giant fucking plasma gas tank outside.
Formerly was only for looks, now can actually inject more plasma into
the tank (no one ever will probably but it makes more sense than having
it just, not hooked up to anything)
-Actually connected plasma tank to Plasma connectors of the gas station.
They weren't. For some reason.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Overall it fixes a blatant mapping error and prevents a Ghost role from
being scorched/burned out so easily. Pretty sure I don't have to
elaborate on the modular part
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
Tested Locally/It's a very minor map edit. I.D still functions the same,
it's only a movement of the code to a modular folder. No new Run-Time
Errors.
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->
<details>
<summary>Screenshots/Videos</summary>


![atmospr1](https://github.com/user-attachments/assets/b1d7649b-12ed-4a07-9c41-c74785ca5b2a)

![image](https://github.com/user-attachments/assets/43710ecd-c13d-4722-98aa-d1015f6b7e98)

p.s the fluid ducting as been moved to connect horizontally instead of
vertically, instead of as defined in the pictures.

</details>



<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->