Lizard Gas Update (#3056)
## About The Pull Request

Following an issue with Miners and other Lavaland/Icemoon roles
ransacking the Gas station of its mats if no attendants spawn in, I've
opted to add blastdoors/shutters, along with warning notes to IC'ly
notify folk that this is a Ghost Role and should be left undisturbed.
Said notes do not use OOC language in doing so, and simply inform the
reader that "hey there are people cryosleeping inside, please dont rob
this place"

Also I added some things I left out in the original overhaul by mistake

-Added Privacy Shutters to all exterior windows, controllable by a
Button in the Break/Spawn Room
-Added Blast-Doors to all exterior Airlocks, controllable by a Button in
the Break/Spawn Room
-Added 2x 'Veteran Lizard Gas Beret' to Locker, a Purple, renamed/desc'd
Badged Beret
-Added 2x 'Lizard Gas Apron' to Locker, a Purple, renamed/desc'd Apron
-Added an Inducer to spawn room, in case late-round joiners spawn in the
Gas Station while it has run out of power.
-Added 2x Insulated Gloves to Locker, in case they need to alter their
doors in case of broken APC/Other reasons the power may be out and not
restorable *(And the doors are still Bolted.)*
-Added Colony Core Replicator Board to `Abandoned Storehouse`. This was
meant to be added to the other flatpacked Colony machines, all it allows
is for clothes/medicine/food to be printed out of Organic material.
-Fixed: The Button in the Manager's Office was not correctly set to Bolt
the door, and instead controlled the Restroom's door. Fixed that.
## Why It's Good For The Game

Keeps people from ransacking the Ghost Role location before it's
spawned/let's people know that it's a Ghost role and shouldn't be
tampered with.

The Locker Additions are primarily quality of life/fluff adds, but the
tools added have little impact/are generally useful to have.
## Proof Of Testing

Purely a minor map change. Ran locally. No additional Run-Time errors.