Replaces The Clown storyteller with The House (Adminbus storyteller) (#2999)
## About The Pull Request

While The House will set weights to one still, it'll be running on the
fragile ruleset (Medium to low chaos)

The catch, admins are given the opportunity to drive the admin bus
straight into the station. The house now runs the show.

## Why It's Good For The Game

I was told by multiple admins that they used Clown as the dedicated
adminbus storyteller. Now I'll make it official, slow down clown, and
give admins the opportunity to take over. The goal is to make it feel
more like a coin flip then just hyper chaos every round.

## Proof Of Testing

works. 

## Changelog

:cl:
add: Replaces The Clown with The House which sets weights to 1, but
spawns stuff at a slower pace, and allows the adminbus to come.
fix: Voidwalker properly spawns on the ghostset now.
/:cl: